feat: Refactor ReportDetailPage to use modular components for AI analysis and original report, improving code organization and readability
