+++ frontend/src/pages/Reports/components/OriginalReportTab.tsx
